Improve XML handling and enhance report status indicators
The MercuryGateAPIservices.rb file was updated to improve XML handling by converting '&' into '&amp;'. A new method was also added into the Static Pages controller to check the existence of different reports. This has been reflected in the views by updating the status indicators of contract and shipment reports by replacing static placeholder with dynamic update based on the availability of the reports. Changes also include modifications in the visual representation of these status indicators.